+++ packages/_private/vue-demo/src/AllComponentParts/DemoButton.vue
@@ -1,13 +1,9 @@
-<script setup lang="ts">
-import { SynVueButton } from '@synergy-design-system/vue';
-</script>
-
 <template>
   <div
     style="display: flex; gap: var(--syn-spacing-medium);"
   >
-    <SynVueButton variant="filled">Filled</SynVueButton>
-    <SynVueButton variant="outline">Outline</SynVueButton>
-    <SynVueButton variant="text">Text</SynVueButton>
+    <syn-button variant="filled">Filled</syn-button>
+    <syn-button variant="outline">Outline</syn-button>
+    <syn-button variant="text">Text</syn-button>
   </div>
 </template>